Add feature to highlight barcodes and text in camera preview
Holding the camera over objects containing recognizable barcodes and/or
text will cause the app to highlight them with a red box. Touching the
red box will either add the barcode to the list or in the case of text,
display it in white text at the bottom of the camera preview.